+++ packages/erxes-ui/src/types.ts
@@ -1,5 +1,5 @@
-import { IActivityLogForMonth } from "@erxes/ui-log/src/activityLogs/types";
-import { IUser } from "./auth/types";
+import { IActivityLogForMonth } from '@erxes/ui-log/src/activityLogs/types';
+import { IUser } from './auth/types';
 
 export interface IRouterProps {
   location: any;
@@ -189,7 +189,7 @@ export type IEditorProps = {
   showMentions?: boolean;
   toolbar?: any[];
   autoFocus?: boolean;
-  toolbarLocation?: "top" | "bottom";
+  toolbarLocation?: 'top' | 'bottom';
   autoGrow?: boolean;
   autoGrowMinHeight?: number | string;
   autoGrowMaxHeight?: number | string;
@@ -200,6 +200,7 @@ export type IEditorProps = {
   additionalToolbarContent?: (props: {
     onClick: (placeholder: string) => void;
   }) => React.ReactNode;
+  contentTypeConfig?: any;
 };
 
 export type QueryResponse = {