+++ packages/plugin-ebarimt-api/src/graphql/resolvers/mutations/index.ts
@@ -1,4 +1,5 @@
 import Ebarimt from './ebarimt';
 import ProductRules from './productRules';
+import ProductGroups from './productGroups';
 
-export { Ebarimt, ProductRules };
+export { Ebarimt, ProductRules, ProductGroups };