+++ packages/ui-tickets/src/boards/components/editForm/Actions.tsx
@@ -51,7 +51,7 @@ class Actions extends React.Component<Props> {
       removeItem,
       sendToBoard,
       onChangeStage,
-      onChangeRefresh
+      onChangeRefresh,
     } = this.props;
 
     const onLabelChange = (labels) => saveItem({ labels });
@@ -132,7 +132,7 @@ class Actions extends React.Component<Props> {
             item,
             contentType: "tickets",
             subType: item.stage?.type,
-            path: `stageId=${item.stageId}`
+            path: `stageId=${item.stageId}`,
           },
           true
         )}